% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fq_round.R
\name{fq.round}
\alias{fq.round}
\title{Redondear frecuencias en MHz}
\usage{
fq.round(fq, bw)
}
\description{
Funcion que redondea una frecuencia en MHz al valor mas proximo segun el
ancho de banda de resolucion en kHz de las mediciones
}
