% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tci_filter_data.R
\name{tci.filter.data}
\alias{tci.filter.data}
\title{Filtrar datos de medicion por frecuencia}
\usage{
tci.filter.data(tci_data, fqrange)
}
\arguments{
\item{tci_data}{Dataframe con el contenido de las mediciones}

\item{fqrange}{Vector con el rango de frecuencias a filtrar}
}
\description{
La funcion permite filtrar por frecuencia los resultados de las mediciones
}
\seealso{
Other TCI file management functions: 
\code{\link{tci.count.skips}()},
\code{\link{tci.filter.index}()},
\code{\link{tci.find.dir}()},
\code{\link{tci.get.data}()},
\code{\link{tci.get.index}()},
\code{\link{tci.import.data}()}
}
\concept{TCI file management functions}
