% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tci_filter_index.R
\name{tci.filter.index}
\alias{tci.filter.index}
\title{Filtrar registros del indice de archivos TCI}
\usage{
tci.filter.index(tci_index, stt_date, stp_date, stt_fq, stp_fq, stn_name, type)
}
\arguments{
\item{tci_index}{Dataframe con informacion de los archivos de medicion TCI
organizados como un indice.}

\item{stt_date}{Fecha inicial del tipo "POSIXct" del rango de datos que se
pretende filtrar}

\item{stp_date}{Fecha final del tipo "POSIXct" del rango de datos}

\item{stt_fq}{Frecuencia inicial en MHz}

\item{stp_fq}{Frecuencia final en MHz}

\item{stn_name}{Nombre de la estacion}

\item{type}{Tipo de archivo a filtrar de acuerdo con el siguiente detalle:
1 = Field Strength vs. Channel,
2 = Field Strength vs. Time of Day,
3 = Occupancy vs. Channel,
4 = Occupancy vs. Time of Day.}
}
\description{
La funcion "tci.filter.index" provee de una herramienta agil para filtrar la
informacion contenida de un indice de archivos de mediciones de TCI. La funcion
devuelve a la salida un vector con datos TRUE en la posicion de aquellos
registros que cumplieron con todos los criterios de busqueda configurados.
}
\examples{
wd <- getwd()
setwd(system.file("extdata", package = "cotfitools"))
start <- as.POSIXct("2021-01-01")
stop <- as.POSIXct("2021-12-31")
mon_info <- tci.find.dir(stt_date = start, stp_date = stop)
n_skips <- tci.count.skips(mon_info)
index <- tci.get.index(mon_info, n_skips)
index_filter <- tci.filter.index(index, stp_date = as.POSIXct("2021-03-01"), type = 1)
View(index[index_filter, ])
index_filter <- tci.filter.index(index, stn_name = "Sonsonate")
View(index[index_filter, ])
index_filter <- tci.filter.index(index, stt_fq = 210, stp_fq = 300)
View(index[index_filter, ])
setwd(wd)
}
\seealso{
Other TCI file management functions: 
\code{\link{graph.occupancy}()},
\code{\link{graph.spectrum}()},
\code{\link{tci.count.skips}()},
\code{\link{tci.filter.data}()},
\code{\link{tci.find.dir}()},
\code{\link{tci.get.data}()},
\code{\link{tci.get.index}()},
\code{\link{tci.import.data}()}
}
\concept{TCI file management functions}
