% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tci_get_data.R
\name{tci.get.data}
\alias{tci.get.data}
\title{Obtener datos de monitoreo con los resultados de la medicion}
\usage{
tci.get.data(index_row, impute_data = FALSE)
}
\arguments{
\item{index_row}{Dataframe de una fila con informacion tipo indice del archivo
a importar. Las columnas de este dataframe de entrada deben estar compuestas
segun lo que resulta de ejecutar la funcion "tci.get.index"}
}
\description{
La funcion permite obtener los datos de monitoreo de un archivo de mediciones
TCI. Para ejecutar la funcion se necesita proporcionar el indice con la
informacion del archivo a importar.
}
\examples{
wd <- getwd()
setwd(system.file("extdata", package = "cotfitools"))
start <- as.POSIXct("2021-01-01")
stop <- as.POSIXct("2021-12-31")
mon_info <- tci.find.dir(stt_date = start, stp_date = stop)
n_skips <- tci.count.skips(mon_info)
index <- tci.get.index(mon_info, n_skips)
View(index)
tci_data <- tci.get.data(index[1, ])
View(tci_Data)
setwd(wd)
}
\seealso{
Other TCI file management functions: 
\code{\link{tci.count.skips}()},
\code{\link{tci.filter.data}()},
\code{\link{tci.filter.index}()},
\code{\link{tci.find.dir}()},
\code{\link{tci.get.index}()},
\code{\link{tci.import.data}()}
}
\concept{TCI file management functions}
