% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tci_get_index.R
\name{tci.get.index}
\alias{tci.get.index}
\title{Build index with TCI measurement task info}
\usage{
get.tci.findex(tci_fdf, nskips)
}
\arguments{
\item{tci_fdf}{Dataframe with the path and the TCI file names to be included
in the index}

\item{nskips}{Numeric vector with details of the number of rows that must be
skipped in each TCI file to extract information.}
}
\description{
The 'get.tci.findex' function allows creating an index with information
concerning the type of parameters that were used to configure each
measurement task.
}
\examples{
wd <- getwd()
setwd(system.file("extdata", package = "cotfitools"))
mon_info <- tci.find.dir()
n_skips <- tci.count.skips(mon_info)
index <- tci.get.index(mon_info, n_skips)
View(index)
setwd(wd)
}
\seealso{
Other TCI file management functions: 
\code{\link{tci.count.skips}()},
\code{\link{tci.filter.data}()},
\code{\link{tci.filter.index}()},
\code{\link{tci.find.dir}()},
\code{\link{tci.get.data}()},
\code{\link{tci.import.data}()}
}
\concept{TCI file management functions}
