% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tci_import_data.R
\name{tci.import.data}
\alias{tci.import.data}
\title{Importar los datos de medicion de un conjunto de archivos de monitoreo}
\usage{
tci.import.data(index_list)
}
\arguments{
\item{index_list}{Dataframe con el indice de la informacion a cargar}
}
\description{
La función "tci.import.data" permite importar los datos de medición de un
conjunto de archivos. Dichos datos son almacenados en un objeto tipo "tibble".
Los datos de cada uno de los archivos importados se almacenan en la variable
'measurements' la cual es acompañada con los datos del indice que ofrecen más
información de las características propias de los datos asociados a estos.
}
\seealso{
Other TCI file management functions: 
\code{\link{tci.count.skips}()},
\code{\link{tci.filter.data}()},
\code{\link{tci.filter.index}()},
\code{\link{tci.find.dir}()},
\code{\link{tci.get.data}()},
\code{\link{tci.get.index}()}
}
\concept{TCI file management functions}
