% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_occupancy.R
\name{graph.occupancy}
\alias{graph.occupancy}
\title{Graficar el porcentaje de ocupacion medido}
\usage{
plot.spectrum(measDF, refLev, DBdiv, Avg, Max)
}
\arguments{
\item{measDF}{Dataframe con la informacion de las mediciones. Dicho dataframe
puede obtenerse por medio de la funcion "tci.get.data".}

\item{Avg}{Si es TRUE se grafica el nivel promedio de campo electrico medido}

\item{Max}{Si es TRUE se grafica el nivel maximo de campo electrico medido}

\item{refLev}{Un numero que representa el nivel de referencia utilizado para
construir el eje Y del grafico}

\item{DBdiv}{longitud de cada recuadro en el eje Y del grafico}

\item{thr}{Nivel umbral utilizado}
}
\description{
La funcion "spectrum.plot" prroporciona una alternativa para construir
graficas que permitan visualizar los niveles de campo electrico medidos en las
tareas de monitoreo
}
\seealso{
Other Graphics functions: 
\code{\link{graph.spectrum}()}
}
\concept{Graphics functions}
