% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_spectrum.R
\name{graph.spectrum}
\alias{graph.spectrum}
\title{Graficar los niveles de campo electrico medidos}
\usage{
spectrum.plot(measDF, refLev, DBdiv, Avg, Max)
}
\arguments{
\item{measDF}{Dataframe con la informacion de las mediciones. Dicho dataframe
puede obtenerse por medio de la funcion "tci.get.data".}

\item{refLev}{Un numero que representa el nivel de referencia utilizado para
construir el eje Y del grafico}

\item{DBdiv}{longitud de cada recuadro en el eje Y del grafico}

\item{Avg}{Si es TRUE se grafica el nivel promedio de campo electrico medido}

\item{Max}{Si es TRUE se grafica el nivel maximo de campo electrico medido}

\item{thr}{Nivel umbral utilizado}
}
\description{
La funcion "spectrum.plot" prroporciona una alternativa para construir
graficas que permitan visualizar los niveles de campo electrico medidos en las
tareas de monitoreo
}
\seealso{
Other Graphics functions: 
\code{\link{graph.occupancy}()}
}
\concept{Graphics functions}
