% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tci_count_skips.R
\name{tci.count.skips}
\alias{tci.count.skips}
\title{Get number of rows until the measurement data}
\usage{
tci.count.skips(tci_fdf)
}
\arguments{
\item{tci_fdf}{Data frame with information of the path and TCI file name to
be read}
}
\description{
The function tci.count.skips identify the number of rows that must
be skipped until measurement data of any TCI file is detected. This
information will permit reading TCI data when used read.csv function since
the number of rows to be skipped is already known.
}
\examples{
wd <- getwd()
setwd(system.file("extdata", package = "cotfitools"))
mon_info <- tci.find.dir()
n_skips <- tci.count.skips(mon_info)
setwd(wd)
}
\seealso{
Other TCI file management functions: 
\code{\link{tci.filter.data}()},
\code{\link{tci.filter.index}()},
\code{\link{tci.find.dir}()},
\code{\link{tci.get.data}()},
\code{\link{tci.get.index}()},
\code{\link{tci.import.data}()}
}
\concept{TCI file management functions}
