% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tci_find_dir.R
\name{tci.find.dir}
\alias{tci.find.dir}
\title{Identify TCI files inside directory}
\usage{
tci.find.dir(__directory path__)
}
\arguments{
\item{main_folder_path}{String of characters with the directory path to be
explored. Working directory is the default path.}
}
\description{
The function tci.find.dir explores the directory path supplied as input
argument and looks for TCI files of measurements data. In brief, the function
navigates inside the given directory path and checks if any file meet
the criteria of a TCI file. As an output, the function brings back a data frame
with the path and the detected TCI file names.
}
\examples{
wd <- getwd()
setwd(system.file("extdata", package = "cotfitools"))
mon_info <- tci.find.dir()
View(mon_info)
setwd(wd)
}
\seealso{
Other TCI file management functions: 
\code{\link{tci.count.skips}()},
\code{\link{tci.filter.data}()},
\code{\link{tci.filter.index}()},
\code{\link{tci.get.data}()},
\code{\link{tci.get.index}()},
\code{\link{tci.import.data}()}
}
\concept{TCI file management functions}
