% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tci_get_index.R
\name{tci.get.index}
\alias{tci.get.index}
\title{Obtener indice con informacion de medicion de archivos TCI}
\usage{
get.tci.findex(tci_fdf, nskips)
}
\arguments{
\item{tci_fdf}{Dataframe con informacion del directorio, nombre del archivo,
fecha inicial y fecha final de la medicion de los archivos TCI.}

\item{nskips}{Vector tipo numerico con el detalle del numero de filas que
deben saltarse en cada archivo TCI para extraer la informacion del indice.}
}
\description{
La funcion 'get.tci.findex' permite crear un indice con infomacion concerniente
al tipo de parametros que se utilizaron para configurar cada tarea de medicion.
}
\examples{
wd <- getwd()
setwd(system.file("extdata", package = "cotfitools"))
start <- as.POSIXct("2021-01-01")
stop <- as.POSIXct("2021-12-31")
mon_info <- tci.find.dir(stt_date = start, stp_date = stop)
n_skips <- tci.count.skips(mon_info)
index <- tci.get.index(mon_info, n_skips)
setwd(wd)
}
\seealso{
Other TCI file management functions: 
\code{\link{graph.occupancy}()},
\code{\link{graph.spectrum}()},
\code{\link{tci.count.skips}()},
\code{\link{tci.filter.data}()},
\code{\link{tci.filter.index}()},
\code{\link{tci.find.dir}()},
\code{\link{tci.get.data}()},
\code{\link{tci.import.data}()}
}
\concept{TCI file management functions}
